Assign Cognitive Impairment Status

Applies cognitive impairment classification based on z-scores computed via the normative regression calculator from Bezdicek et al. (2017), following Level II PD-MCI criteria proposed by Litvan et al. (2012). A patient is classified as cognitively impaired if they have two or more test scores with z < -1.5.

Usage

assign_cognitive_impairment(d0, calc, map, output = "both")

Arguments

d0

A list of data, typically generated by prepare_data, containing test scores.

calc

A character string specifying the file path to the regression calculator (.xlsx format).

map

A data frame, tibble, or matrix with:

Column 1: Label used in the calculator file

Column 2: Corresponding label in the data

Column 3: Cognitive domain of the test

Column 4: Indicator of score direction (1 for normal, -1 if reversed)

Alternatively, a semicolon-delimited CSV file path containing such a table.

output

A character string specifying the output format:

"ci" – return only a logical indicator of cognitive impairment

"z" – return only z-scores for each index

"both" – return both (default)

Value

Depending on the output argument:

"ci"

A logical vector indicating cognitive impairment (TRUE/FALSE).

"z"

A data frame of computed z-scores per cognitive test.

"both"

A data.frame containing both z (z-scores) and ci (impairment status).

References

Bezdicek, O., Sulc, Z., Nikolai, T., Stepankova, H., Kopecek, M., Jech, R., & Růžička, E. (2017). A parsimonious scoring and normative calculator for the Parkinson’s disease mild cognitive impairment battery. The Clinical Neuropsychologist, 31(6-7), 1231-1247. doi:10.1080/13854046.2017.1293161

Litvan, I., Goldman, J. G., Tröster, A. I., Schmand, B. A., Weintraub, D., Petersen, R. C., ... & Emre, M. (2012). Diagnostic criteria for mild cognitive impairment in Parkinson's disease: Movement Disorder Society Task Force guidelines. Movement disorders, 27(3), 349-356. doi:10.1002/mds.24893

See also

Examples

if (FALSE) { # \dontrun{
ci <- assign_cognitive_impairment(
  data,
  here::here("data-raw", "LevelIICalculator.xlsx"),
  here::here("data-raw", "CalculatorMapping.csv")
)
} # }