Make a Dendrogram of PDD Algorithms

The functions prepares a dendrogram (tree plot) that summarises PDD algorithms construction starting from IADL operationalisation at the top followed successively by operationalisation of Global cognitive performance, Attention, Executive Function, Construction, Memory and Language. The dendrogram visualises PDD rate estimated by each algorithm via point size.

Usage

make_dendrogram(tab, top)

Arguments

tab

A table containing all criteria used as generated by summarise_rates.

top

A tibble containing top criteria for predicting PDD diagnosis by Level II a generated by table_levelII_approximations.

Value

A ggplot object with a dendrogram created by ggraph.

See also

Examples

if (FALSE) { # \dontrun{
# Raw data:
p <- data_paths("data-raw")
data <- prepare_data(p)
pdd  <- diagnose_pdd_sample(data)

# PDD rates for point size:
vars <- here::here("data-raw", "VariablesOfInterest.csv")
rates <- summarise_rates(pdd, vars)

# Concordance with Level II for edge width:
algos <- list_algorithms(rates$table)
data(concords)
best <- table_levelII_approximations(concords, algos, "Accuracy_raw")

tree <- make_dendrogram(rates$table, best$table)
} # }