Perform variable selection

A wrapper for projpred::cv_varsel(). Using a provided reference model, runs a projective prediction feature selection via cross-validation.

Usage

perform_varsel(refm_fit, ...)

Arguments

refm_fit

Reference model, e.g., fitted via fit_reference.

...

Parameters passed to projpred::cv_varsel().

Value

An object of class vsel as returned by projpred::cv_varsel().

See also