Plot correlation matrixes

Generates two correlation matrixes via corrplot::corrplot() and plots them side to side. Intended to be used for visual comparison of observed and "ideal" correlations as generated by correlate_indexes. Format is fixed for purposes of {demcrit}.

Usage

plot_correlation(corrs_obs, corrs_opt = NULL)

Arguments

corrs_obs

Observed correlation matrix.

corrs_opt

Ideal correlation matrix. If NULL, ontly the observed matrix will be plotted.

Value

Plots the correlation matrixes from corrs side by side; with left (A) depicting observed correlations, right (B) depicting the "ideal" correlation matrix; or a single correlation matrix.